% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cc-palette.R
\docType{data}
\name{cc_palette}
\alias{cc_palette}
\title{CC corporate colour palette}
\format{
Length 8 named vector
}
\usage{
cc_palette
}
\description{
There seems to be some confusion over students 1 and 2, and partners 1 and 2.
My understanding is that v1 are the new ones
}
\examples{

scales::show_col(cc_palette)

}
\keyword{datasets}
