% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student-repos.R
\name{clone_hw_repos}
\alias{clone_hw_repos}
\title{Clone student repositories}
\usage{
clone_hw_repos(form, ..., hw_root = "~/Documents")
}
\arguments{
\item{form}{Dataframe; Class \code{hw_form}}

\item{...}{Character; any sub-directories to add onto hw_root}

\item{hw_root}{Character; Root path to where repositories should be cloned}
}
\value{
Dataframe of two columns: name, and github url (invisible)
}
\description{
Extract student names and Github repo links from the homework submission form
found on Google Sheets. Repositories are cloned into the directory given by
\code{hw_root}.
}
\examples{
\dontrun{
 clone_hw_repos(hw_form, 'homework', 'de999')
}

}
