% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student-repos.R
\name{dl_hw_form}
\alias{dl_hw_form}
\title{Download homework submission form}
\usage{
dl_hw_form(g_sheet, ...)
}
\arguments{
\item{g_sheet}{Character; Google Sheet identifier, see
\code{\link[googlesheets4:read_sheet]{read_sheet}} for details.}

\item{...}{Does nowt (currently)}
}
\value{
Data frame
}
\description{
Download a GoogleSheets spreadsheet from the given link. The columns are
renamed to make them less unwieldy in a data frame. The original names are
preserved as an attribute.
}
\details{
The first time this is called, you should be taken to the Google sign in
page, where it will be necessary to enter authenticate. RStudio should then
offer to allow the credentials to persist across sessions - your choice.
}
\examples{
\dontrun{
gs_url <- 'https://docs.google.com/spreadsheets/d/...unique google sheets url...'
form <- dl_hw_form(gs_url)
}

}
