% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open-notes-files.R
\name{find_notes}
\alias{find_notes}
\title{Locate files at given path}
\usage{
find_notes(
  week,
  day,
  file_types = c("notes", "data", "rmd"),
  path,
  drop_dups = TRUE
)
}
\arguments{
\item{week}{Integer}

\item{day}{Integer}

\item{file_types}{Character; Any or all of 'notes', 'rmd', or 'data'}

\item{path}{Character; Location of course repo}

\item{drop_dups}{Logical; Remove duplicated dataset names}
}
\value{
List of character vector filepaths named 'notes', 'rmd' and 'data'
}
\description{
Determines all notes files (html, Rmd) or data (csv) files at the given location.
}
\details{
If \code{file_types} is \code{'notes'}, html note files are opened. If \code{file_types} is \code{'rmd'} Rmarkdown notebooks are opened. If \code{file_types} is \code{'data'} csv files are read and placed in the global environment. Data files are named according to their \code{\link{basename}}
}
