% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open-notes-files.R
\name{open_notes}
\alias{open_notes}
\title{Opens notes files}
\usage{
open_notes(
  week,
  day,
  file_types = c("notes", "data", "rmd"),
  path = "/Users/user/Documents/GitHub/data_rewrite2"
)
}
\arguments{
\item{week}{Integer}

\item{day}{Integer}

\item{file_types}{Character; Any or all of 'notes', 'rmd', or 'data'}

\item{path}{Character; Location of course repo}
}
\description{
Loops over a vector or file paths and open the files indicated.
}
