% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student-repos.R
\name{pull_hw}
\alias{pull_hw}
\title{Pull to student repos}
\usage{
pull_hw(hw_dir, cohort = "")
}
\arguments{
\item{hw_dir}{Character; Path to homework directory}

\item{cohort}{Character; Cohort directory}
}
\value{
List, named by student, containing information on the result of the
  pull operation. This is returned invisibly, so if it is required, it should
  be assigned
}
\description{
Update student homework repositories.
}
\details{
The returned result is a list of [ git response stuff that i should probably
detail more precisely ], named according to the CodeClan naming convention
for student homework repositories. The path provided to \code{hw_dir} can be
the path to a single repository, or a directory containing one or more
repositories. In the latter case, the \code{cohort} parameter may also be
used to specify a sub-directory.
}
\examples{
\dontrun{
  hw_root_dir <- '/Users/user/Documents/GitHub/homework'
  cohort <- 'de999'
  pull_response <- pull_hw(c('zzz', hw_root_dir), cohort) # 'zzz' fails
}

}
