% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/student-repos.R
\name{setup_cohort_repos}
\alias{setup_cohort_repos}
\title{Setup student repo clones}
\usage{
setup_cohort_repos(gs_url, ...)
}
\arguments{
\item{gs_url}{}
}
\value{
Dataframe of two columns: name, and github url (invisible)
}
\description{
NEED TO TEST. Is just a wrapper around download & clone functions
}
