% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open-notes-files.R
\name{open_notes}
\alias{open_notes}
\title{Opens notes files}
\usage{
open_notes(
  week,
  day,
  file_types = c("notes", "data", "rmd"),
  path = "/Users/user/Documents/GitHub/data_rewrite2",
  clean_names = TRUE,
  drop_dups = TRUE
)
}
\arguments{
\item{week}{Integer}

\item{day}{Integer}

\item{file_types}{Character; Any or all of 'notes', 'rmd', or 'data'}

\item{path}{Character; Location of course repo}

\item{clean_names}{Logical; Pass dataset columns names through \code{\link{janitor::clean_names}}}

\item{drop_dups}{Logical; Remove duplicated dataset names}
}
\value{
Character; names of datasets attached.
}
\description{
Loops over a vector or file paths and open the files indicated. If two or
more datasets have the same name basename, only the last read will be
available
}
